requirements.txt
# Data fetching
euroleague-api==0.1.0
requests==2.32.3
python-dotenv==1.0.1

# Data processing
pandas==2.2.3
numpy==2.1.3

# Google Sheets integration
gspread==6.1.4
google-auth==2.37.0
google-auth-oauthlib==1.2.1

# Utilities
tqdm==4.67.1
colorlog==6.9.0
